fix for JBEHAVE-813 and JBEHAVE-815
JBEHAVE-813: html template report contains encoded html tags

correctly encode step pattern and parameters but not the html/xml tags that are inserted into the text in the report
(changes for the actual fix and added a few lines to the TemplateOutputBehavious as well as to the PrintStreamOutputBehaviour, this does the encoding correctly however)
To distinguish between XML and HTML I had to introduce an enum that is accessible from the freemarker template, which is kind of complicated.

JBEHAVE-815: template reports do not mark step parameter if it contains special chars

the parameters are matched by a regex including specific chars not including e.g. angle brackets, comma etc, this is unnecessary if the regex is matched non-greedy.